+++ src/bflib_network.cpp
@@ -20,7 +20,7 @@
 #include "bflib_network.h"
 
 #include "bflib_basics.h"
-#include "bflib_datetm.h"
+#include "bflib_datetm.hpp"
 #include "bflib_memory.h"
 #include "bflib_netsession.h"
 #include "bflib_netsp.hpp"
