+++ src/bflib_pom.cpp
@@ -22,7 +22,7 @@
 #include <time.h>
 
 #include "bflib_pom.hpp"
-#include "bflib_datetm.h"
+#include "bflib_datetm.hpp"
 
 using namespace std;
 
@@ -206,7 +206,7 @@ long PhaseOfMoon::JulToday(double &daypart,long &moon_periods_n)
 
 /**
  * Library exported routine for calculating Phase Of Moon.
- * Defined in Ansi-C header "bflib_datetm.h".
+ * Defined in Ansi-C header "bflib_datetm.hpp".
  */
 double LbMoonPhase(void)
 {
