+++ src/front_input.c
@@ -26,7 +26,7 @@
 #include "bflib_keybrd.h"
 #include "bflib_mouse.h"
 #include "bflib_sprfnt.h"
-#include "bflib_datetm.h"
+#include "bflib_datetm.hpp"
 #include "bflib_fileio.h"
 #include "bflib_memory.h"
 #include "bflib_network.h"
