+++ src/front_simple.c
@@ -23,7 +23,7 @@
 #include "bflib_memory.h"
 #include "bflib_keybrd.h"
 #include "bflib_inputctrl.h"
-#include "bflib_datetm.h"
+#include "bflib_datetm.hpp"
 #include "bflib_video.h"
 #include "bflib_fileio.h"
 #include "bflib_dernc.h"
