+++ src/frontend.cpp
@@ -26,7 +26,7 @@
 #include "bflib_sprite.h"
 #include "bflib_sprfnt.h"
 #include "bflib_dernc.h"
-#include "bflib_datetm.h"
+#include "bflib_datetm.hpp"
 #include "bflib_keybrd.h"
 #include "bflib_inputctrl.h"
 #include "bflib_sndlib.h"
