+++ src/frontmenu_ingame_evnt.c
@@ -19,7 +19,7 @@
 #include "frontmenu_ingame_evnt.h"
 #include "globals.h"
 #include "bflib_basics.h"
-
+#include "bflib_datetm.hpp"
 #include "bflib_guibtns.h"
 #include "bflib_vidraw.h"
 #include "bflib_sprfnt.h"
@@ -44,8 +44,6 @@
 #include "keeperfx.hpp"
 
 unsigned long TimerTurns = 0;
-
-int debug_display_frametime = 0;
 
 /******************************************************************************/
 void gui_open_event(struct GuiButton *gbtn)
@@ -554,12 +552,16 @@ void draw_frametime()
     LbTextSetFont(winfont);
     lbDisplay.DrawFlags = Lb_TEXT_HALIGN_RIGHT;
     int tx_units_per_px = (11 * units_per_pixel) / LbTextLineHeight();
+
+    // FPS
+    display_value = 1000 / frametime_measurements.frametime_display[Frametime_FullFrame];
+
+    text = buf_sprintf("FPS: %f", display_value);
+    LbTextDrawResized(0, 27*tx_units_per_px, tx_units_per_px, text);
+
+    // Frametimes
     for (int i = 0; i < TOTAL_FRAMETIME_KINDS; i++) {
-        if (debug_display_frametime == 1) {
-            display_value = frametime_measurements.frametime_display[i];
-        } else {
-            display_value = frametime_measurements.frametime_display_max[i];
-        }
+        display_value = frametime_measurements.frametime_display[i];
         switch (i) {
             case Frametime_FullFrame:
                 text = buf_sprintf("Frametime: %f ms", display_value);
