+++ src/frontmenu_net.c
@@ -20,7 +20,7 @@
 #include "globals.h"
 #include "bflib_basics.h"
 
-#include "bflib_datetm.h"
+#include "bflib_datetm.hpp"
 #include "bflib_guibtns.h"
 #include "bflib_video.h"
 #include "bflib_vidraw.h"
