+++ src/gui_frontbtns.c
@@ -24,7 +24,7 @@
 #include "bflib_vidraw.h"
 #include "bflib_sprite.h"
 #include "bflib_sprfnt.h"
-#include "bflib_datetm.h"
+#include "bflib_datetm.hpp"
 #include "kjm_input.h"
 #include "gui_draw.h"
 #include "gui_frontmenu.h"
