+++ src/main_game.c
@@ -1,7 +1,7 @@
 #include "keeperfx.hpp"
 
 #include "bflib_coroutine.h"
-#include "bflib_datetm.h"
+#include "bflib_datetm.hpp"
 #include "bflib_math.h"
 #include "bflib_memory.h"
 #include "bflib_sound.h"
