+++ src/player_instances.c
@@ -20,7 +20,7 @@
 
 #include "globals.h"
 #include "bflib_basics.h"
-#include "bflib_datetm.h"
+#include "bflib_datetm.hpp"
 #include "bflib_math.h"
 #include "bflib_sound.h"
 #include "bflib_planar.h"
